restructure and refactor
