+++ cmd/kai/main.go
@@ -9,6 +9,7 @@ import (
 	"syscall"
 
 	"github.com/basebandit/kai"
+	"github.com/basebandit/kai/clustermanager"
 	"github.com/basebandit/kai/tools"
 )
 
@@ -16,7 +17,7 @@ func main() {
 	kubeconfig := filepath.Join(
 		os.Getenv("HOME"), ".kube", "config",
 	)
-	cm := kai.NewClusterManager()
+	cm := clustermanager.New()
 
 	err := cm.LoadKubeConfig("local", kubeconfig)
 	if err != nil {
@@ -47,7 +48,7 @@ func main() {
 	fmt.Fprintln(os.Stdout, "Server stopped")
 }
 
-func registerAllTools(s *kai.Server, cm *kai.ClusterManager) {
+func registerAllTools(s *kai.Server, cm *clustermanager.Cluster) {
 	tools.RegisterPodTools(s, cm)
 	tools.RegisterDeploymentTools(s, cm)
 }
